Split and cleanup integration tests
