+++ phpunit.xml.dist
@@ -15,6 +15,6 @@
         </testsuite>
     </testsuites>
     <php>
-        <env name="STRIPE_MODEL" value="Laravel\Cashier\Tests\Integration\User"/>
+        <env name="STRIPE_MODEL" value="Laravel\Cashier\Tests\Fixtures\User"/>
     </php>
 </phpunit>
