+++ tests/Integration/IntegrationTestCase.php
@@ -1,0 +1,117 @@
+<?php
+
+namespace Laravel\Cashier\Tests\Integration;
+
+use Stripe\Token;
+use Stripe\Stripe;
+use Stripe\ApiResource;
+use PHPUnit\Framework\TestCase;
+use Stripe\Error\InvalidRequest;
+use Illuminate\Database\Schema\Builder;
+use Laravel\Cashier\Tests\Fixtures\User;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\ConnectionInterface;
+use Illuminate\Database\Capsule\Manager as DB;
+use Illuminate\Database\Eloquent\Model as Eloquent;
+
+abstract class IntegrationTestCase extends TestCase
+{
+    /**
+     * @var string
+     */
+    protected static $stripePrefix = 'cashier-test-';
+
+    public static function setUpBeforeClass()
+    {
+        parent::setUpBeforeClass();
+
+        Stripe::setApiKey(getenv('STRIPE_SECRET'));
+    }
+
+    public function setUp()
+    {
+        parent::setUp();
+
+        Eloquent::unguard();
+
+        $db = new DB;
+        $db->addConnection([
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+        ]);
+        $db->bootEloquent();
+        $db->setAsGlobal();
+
+        $this->schema()->create('users', function (Blueprint $table) {
+            $table->increments('id');
+            $table->string('email');
+            $table->string('name');
+            $table->string('stripe_id')->nullable();
+            $table->string('card_brand')->nullable();
+            $table->string('card_last_four')->nullable();
+            $table->timestamps();
+        });
+
+        $this->schema()->create('subscriptions', function (Blueprint $table) {
+            $table->increments('id');
+            $table->integer('user_id');
+            $table->string('name');
+            $table->string('stripe_id');
+            $table->string('stripe_plan');
+            $table->integer('quantity');
+            $table->timestamp('trial_ends_at')->nullable();
+            $table->timestamp('ends_at')->nullable();
+            $table->timestamps();
+        });
+    }
+
+    public function tearDown()
+    {
+        $this->schema()->drop('users');
+        $this->schema()->drop('subscriptions');
+    }
+
+    protected static function deleteStripeResource(ApiResource $resource)
+    {
+        try {
+            $resource->delete();
+        } catch (InvalidRequest $e) {
+            //
+        }
+    }
+
+    protected function schema(): Builder
+    {
+        return $this->connection()->getSchemaBuilder();
+    }
+
+    protected function connection(): ConnectionInterface
+    {
+        return Eloquent::getConnectionResolver()->connection();
+    }
+
+    protected function getTestToken($cardNumber = null)
+    {
+        return Token::create([
+            'card' => [
+                'number' => $cardNumber ?? '4242424242424242',
+                'exp_month' => 5,
+                'exp_year' => date('Y') + 1,
+                'cvc' => '123',
+            ],
+        ])->id;
+    }
+
+    protected function getInvalidCardToken()
+    {
+        return $this->getTestToken('4000 0000 0000 0341');
+    }
+
+    protected function createCustomer($description = 'taylor'): User
+    {
+        return User::create([
+            'email' => "{$description}@cashier-test.com",
+            'name' => 'Taylor Otwell',
+        ]);
+    }
+}
